4.2.5 add private method in task and task-edit
